Add a node.js script to (re-)import content from explainXKCD
The wonderful site at https://www.explainxkcd.com/ was clearly the
source of the content stored in this here repository.

But it has not been synchronized in quite a while, which is too bad
because XKCD is this wonderful source of joy and being able to quickly
search it is a very, very nice thing.

So let's add a somewhat simplistic node.js script that takes on its
standard input an XML dump of the most recent revisions of the pages on
that site.

The idea is to remove the files in `content/comics/` and then run this
script to re-generate them.

Note that this script does not by any form or shape claim to be a
complete converter from Wikitext to Markdown! It is _just_ good enough
for the task at hand, and if gaps are found in the conversion, the idea
is to adapt the script accordingly and re-import the pages.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>