Add a scheduled GitHub workflow to re-import from explainXKCD
I already have a script that imports the comics from an XML dump that is
exported from explainXKCD.

Sadly, I have not found an easy way to generate such a dump via the
official Mediawiki API, not without jumping through lots of hassles.

But there is this wonderful project called Playwright
(https://playwright.dev/) that allows remote-controlling browsers. Its
official purpose is to implement UI tests for websites, but I can just
as well use it to repeat those otherwise-manual steps of downloading
that XML dump.

This allows me to add a GitHub workflow that updates the comics on a
regular schedule in a fully-automated fashion.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>